Clean ups:
std_machine/accelerator: failed test to silence warnings fpr
                         casting integers
example/lattice_test: added check that lattice file was given

core/field_interpolation/multiples: need to understand how to handle grqdient

elements/classical_magnet: check that interpolator is registered
       rather a debug check

tps: improved comments. checks for exporting ss_vect<tps>
     e.g. filling exported arma matrix with NAN's

python: added factorý for creating accelerator from configuration file
        now used by a_machine and observe_pass examples
        exporting ss_vect<tps> as arma matrix. Simple example added
        for illustrating functionality. Observer using
        this matrix to store tps values

python now exporting armadillo martices using buffer protocol
thus find_packae move to root dir CMakeLists

+++ src/thor_scsi/std_machine/accelerator.cc
@@ -13,9 +13,9 @@ void
 ts::Accelerator::_propagate(thor_scsi::core::ConfigType& conf, ss_vect<T> &ps, size_t start, int max)// const
 {
 
-	const int nelem = int(this->size());
+	const int nelem = static_cast<int>(this->size());
 
-	int next_elem = start;
+	int next_elem = static_cast<int>(start);
 	bool retreat = std::signbit(max);
 
 	for(int i=0; next_elem >= 0 && next_elem<nelem && i<abs(max); i++)
